[refactor] config function lookup moved to CLI
