+++ hypergan/gan.py
@@ -41,30 +41,6 @@ class GAN:
     def sample_file(self, name, sampler=grid_sampler):
         sampler.sample(name, self.sess, self.config)
 
-    # This looks up a function by name.   Should it be part of hyperchamber?
-    #TODO moveme
-    def get_function(self, name):
-        if name == "function:hypergan.util.ops.prelu_internal":
-            return prelu("g_")
-
-        if not isinstance(name, str):
-            return name
-        namespaced_method = name.split(":")[1]
-        method = namespaced_method.split(".")[-1]
-        namespace = ".".join(namespaced_method.split(".")[0:-1])
-        return getattr(importlib.import_module(namespace),method)
-
-    # Take a config and replace any string starting with 'function:' with a function lookup.
-    #TODO moveme
-    def lookup_functions(self, config):
-        for key, value in config.items():
-            if(isinstance(value, str) and value.startswith("function:")):
-                config[key]=self.get_function(value)
-            if(isinstance(value, list) and len(value) > 0 and isinstance(value[0],str) and value[0].startswith("function:")):
-                config[key]=[self.get_function(v) for v in value]
-
-        return config
-
     def create_graph(self, x, y, f, graph_type, device):
         self.graph = hg.graph.Graph(self.config)
